Fix dingding moving
